Simplify test structure
